add new theme

+++ src/helpers.php
@@ -87,36 +87,24 @@ function activate_my_theme() {
     $template_directory= get_template_directory();
     $dir_parts_blocks    = $template_directory.'/acf-blocks/';
     $dir_parts_src    = $template_directory.'/assets/src/';
-    $folders = scandir($dir_parts_blocks);
-    $text_scss = "@import 'app';\r\n";
-    $text_js = "import import_js from './app.js';\r\nimport '../scss/styles.scss';\r\n";
-    foreach($folders as $folder){
-        if($folder=='..'||$folder=='.'||$folder=='_import.scss'||$folder=='import.js'){
-            continue;
-        }
-        $text_scss.="@import './../../../acf-blocks/".$folder."/scss/app';\r\n";
-        $text_js.="import ".preg_replace('/[^ a-zа-яё\d]/ui', '',$folder )."_block from '../../../acf-blocks/".$folder."/js/app.js';\r\n";
-    };
+    if(file_exists($dir_parts_src)&&file_exists($dir_parts_blocks)){
+        $folders = scandir($dir_parts_blocks);
+        $text_scss = "@import 'app';\r\n";
+        $text_js = "import import_js from './app.js';\r\nimport '../scss/styles.scss';\r\n";
+        foreach($folders as $folder){
+            if($folder=='..'||$folder=='.'||$folder=='_import.scss'||$folder=='import.js'||$folder=='.DS_Store'){
+                continue;
+            }
+            $text_scss.="@import './../../../acf-blocks/".$folder."/scss/app';\r\n";
+            $text_js.="import ".preg_replace('/[^ a-zа-яё\d]/ui', '',$folder )."_block from '../../../acf-blocks/".$folder."/js/app.js';\r\n";
+        };
 //    file_put_contents($dir_parts_blocks.'_import.scss', $text_scss);
 //    file_put_contents($dir_parts_blocks.'import.js', $text_js);
-    file_put_contents($dir_parts_src.'scss/styles.scss', $text_scss);
-    file_put_contents($dir_parts_src.'js/script.js', $text_js);
-}
-
-/**
- * Verify for the presence of real content in the standard WordPress editor
- *
- * @param $content
- *
- * @return string
- */
-function getContentWpEditor($content):string
-{
-    $pattern = ['<p>', '</p>', '&nbsp;', '<br>', '</br>', '<br />'];
-
-    if ( trim(str_replace($pattern, '', $content)) ) {
-        return $content;
-    } else {
-        return '';
+        if(file_exists($dir_parts_src.'scss/styles.scss')){
+            file_put_contents($dir_parts_src.'scss/styles.scss', $text_scss);
+        }
+        if(file_exists($dir_parts_src.'js/script.js')){
+            file_put_contents($dir_parts_src.'js/script.js', $text_js);
+        }
     }
 }
